requirements.txt
fastapi
uvicorn
groq
websockets
python-dotenv
psycopg2-binary
